Add JSON Printing Plates
By making a book filled with JSON Raw Text and the running the recipe,
it'll take the raw text, sanitize it, and then output a printing plate
that can be used as normal to print very fancy books.

To have multiple pages, you simply insert the string <<PAGE>> to split off a
new page.